+++ src/GridCal/Gui/Main/ConsoleLog.py
@@ -3,7 +3,7 @@
 # Form implementation generated from reading ui file 'ConsoleLog.ui',
 # licensing of 'ConsoleLog.ui' applies.
 #
-# Created: Mon Jul  6 21:21:47 2020
+# Created: Fri Jul 17 10:56:42 2020
 #      by: pyside2-uic  running on PySide2 5.13.0
 #
 # WARNING! All changes made in this file will be lost!